+++ pom.xml
@@ -136,6 +136,7 @@
         <mysql.jdbc.version>8.0.30</mysql.jdbc.version>
 
         <micrometer-core.version>1.8.0</micrometer-core.version>
+        <io.kubernetes.client-java>18.0.0</io.kubernetes.client-java>
     </properties>
 
     <dependencyManagement>
@@ -1084,7 +1085,7 @@
             <dependency>
                 <groupId>io.kubernetes</groupId>
                 <artifactId>client-java</artifactId>
-                <version>18.0.0</version>
+                <version>${io.kubernetes.client-java}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
