+++ server/odc-service/src/main/java/com/oceanbase/odc/service/task/caller/BasePrimitiveK8SJobClient.java
@@ -29,9 +29,9 @@ import io.kubernetes.client.util.Config;
  * @date 2023-11-15
  * @since 4.2.4
  */
-public abstract class BasePrimitiveK8sClient implements K8sClient {
+public abstract class BasePrimitiveK8SJobClient implements K8sJobClient {
 
-    public BasePrimitiveK8sClient(String k8sClusterUrl) throws IOException {
+    public BasePrimitiveK8SJobClient(String k8sClusterUrl) throws IOException {
         ApiClient apiClient = Config.defaultClient();
         apiClient.setHttpClient(apiClient
                 .getHttpClient()
@@ -42,9 +42,5 @@ public abstract class BasePrimitiveK8sClient implements K8sClient {
                 .setBasePath(k8sClusterUrl);
         Configuration.setDefaultApiClient(apiClient);
     }
-
-    protected abstract String getVersion();
-
-    protected abstract String getKind();
 
 }
