+++ server/odc-service/src/test/java/com/oceanbase/odc/service/k8s/PrimitiveK8sClientTest.java
@@ -28,9 +28,9 @@ import org.junit.Test;
 
 import com.oceanbase.odc.service.task.caller.JobException;
 import com.oceanbase.odc.service.task.caller.JobUtils;
-import com.oceanbase.odc.service.task.caller.K8sClient;
+import com.oceanbase.odc.service.task.caller.K8sJobClient;
 import com.oceanbase.odc.service.task.caller.PodParam;
-import com.oceanbase.odc.service.task.caller.PrimitiveJobBasedK8sClient;
+import com.oceanbase.odc.service.task.caller.PrimitiveK8sJobClient;
 import com.oceanbase.odc.test.database.TestProperties;
 
 /**
@@ -39,19 +39,19 @@ import com.oceanbase.odc.test.database.TestProperties;
  * @since 4.2.4
  */
 @Ignore("manual test this case because k8s cluster is not public environment")
-public class PrimitiveJobBasedK8sClientTest {
+public class PrimitiveK8sClientTest {
 
-    private static K8sClient k8sClient;
+    private static K8sJobClient k8sClient;
 
     @BeforeClass
     public static void init() throws IOException {
-        k8sClient = new PrimitiveJobBasedK8sClient(TestProperties.getProperty("odc.k8s.cluster.url"));
+        k8sClient = new PrimitiveK8sJobClient(TestProperties.getProperty("odc.k8s.cluster.url"));
     }
 
     @Test
     public void test_createJob() throws JobException {
 
-        long taskId = 1L;
+        long taskId = System.currentTimeMillis();
         String imageName = "perl:5.34.0";
         String exceptedJobName = JobUtils.generateJobName(taskId);
         List<String> cmd = Arrays.asList("perl", "-Mbignum=bpi", "-wle", "print bpi(2000)");
@@ -63,6 +63,9 @@ public class PrimitiveJobBasedK8sClientTest {
         Optional<String> queryJobName = k8sClient.getNamespaceJob("default", exceptedJobName);
         Assert.assertTrue(queryJobName.isPresent());
         Assert.assertEquals(exceptedJobName, queryJobName.get());
+
+        String deleteJobOfName = k8sClient.deleteNamespaceJob("default", exceptedJobName);
+        Assert.assertEquals(exceptedJobName, deleteJobOfName);
     }
